Converted setDutyCycle to use uint32_t. Signal is ActiveHigh more often however

+++ src/io/platform/f3xx/PWMf3xx.cpp
@@ -220,7 +220,7 @@ PWMf3xx::PWMf3xx(Pin pin) : PWM(pin) {
                             alternateFunction);
 }
 
-void PWMf3xx::setDutyCycle(float dutyCycle) {
+void PWMf3xx::setDutyCycle(uint32_t dutyCycle) {
     this->dutyCycle = dutyCycle;
 
     TIM_OC_InitTypeDef sConfigOC = {0};
